Renaming all tasks, moving to sub-folders, and starting to create /dest process.
