+++ tasks/javascript/jsTestCombine.js
@@ -2,7 +2,7 @@
 module.exports = function(gulp, sequence) {
     return function() {
         'use strict';
-        sequence('eslint', 'jsbuild', 'jstestbuild', 'jstest', 'browsersyncreload');
+        sequence('jsTestBuild', 'jsTest', 'browsersyncReload');
     };
 };
